Fixed bug in ngrok API response schema

+++ src/NgrokExtensionsSolution/NgrokExtensions/NgrokTunnelsApiResponse.cs
@@ -37,25 +37,25 @@ namespace NgrokExtensions
     {
         public int count { get; set; }
         public int gauge { get; set; }
-        public int rate1 { get; set; }
-        public int rate5 { get; set; }
-        public int rate15 { get; set; }
-        public int p50 { get; set; }
-        public int p90 { get; set; }
-        public int p95 { get; set; }
-        public int p99 { get; set; }
+        public decimal rate1 { get; set; }
+        public decimal rate5 { get; set; }
+        public decimal rate15 { get; set; }
+        public decimal p50 { get; set; }
+        public decimal p90 { get; set; }
+        public decimal p95 { get; set; }
+        public decimal p99 { get; set; }
     }
 
     public class Http
     {
         public int count { get; set; }
-        public int rate1 { get; set; }
-        public int rate5 { get; set; }
-        public int rate15 { get; set; }
-        public int p50 { get; set; }
-        public int p90 { get; set; }
-        public int p95 { get; set; }
-        public int p99 { get; set; }
+        public decimal rate1 { get; set; }
+        public decimal rate5 { get; set; }
+        public decimal rate15 { get; set; }
+        public decimal p50 { get; set; }
+        public decimal p90 { get; set; }
+        public decimal p95 { get; set; }
+        public decimal p99 { get; set; }
     }
 
 }
